Loki: Rename public/app/plugins/datasource/loki/language_provider.ts file to follow convention
