+++ public/app/plugins/datasource/loki/components/LokiLabelBrowser.tsx
@@ -18,7 +18,7 @@ import {
 } from '@grafana/ui';
 
 import PromQlLanguageProvider from '../../prometheus/language_provider';
-import LokiLanguageProvider from '../language_provider';
+import LokiLanguageProvider from '../LanguageProvider';
 import { escapeLabelValueInExactSelector, escapeLabelValueInRegexSelector } from '../language_utils';
 
 // Hard limit on labels to render
