+++ public/app/plugins/datasource/prometheus/promql.ts
@@ -3,7 +3,7 @@ import { Grammar } from 'prismjs';
 import { CompletionItem } from '@grafana/ui';
 
 // When changing RATE_RANGES, check if Loki/LogQL ranges should be changed too
-// @see public/app/plugins/datasource/loki/language_provider.ts
+// @see public/app/plugins/datasource/loki/LanguageProvider.ts
 export const RATE_RANGES: CompletionItem[] = [
   { label: '$__interval', sortValue: '$__interval' },
   { label: '$__rate_interval', sortValue: '$__rate_interval' },
